push Count

+++ ProgramMath/Algebra.cs
@@ -18,8 +18,8 @@ namespace ProgramMath
         public float A { get; set; }
         public float B { get; set; }
         public float C { get; set; }
-        public double Delta => B*B - 4*A*C;
-        public double X => (-B) / (2 * A);
+        public float Delta => B*B - 4*A*C;
+        public float X => (-B) / (2 * A);
         public double X1 => (-B + Math.Sqrt(Delta)) / (2 * A);
         public double X2 => (-B - Math.Sqrt(Delta)) / (2 * A); 
     }
